Extract HeroSection, StatCards, NextActions, and LeadCoverage into separate component files
